+++ docs/demos/theme_explorer/list_group.py
@@ -1,0 +1,24 @@
+import dash_bootstrap_components as dbc
+import dash_html_components as html
+
+from .util import make_subheading
+
+list_group = html.Div(
+    [
+        make_subheading("ListGroup", "list_group"),
+        dbc.ListGroup(
+            [
+                dbc.ListGroupItem("Item 1", color="primary", action=True),
+                dbc.ListGroupItem("Item 2"),
+                dbc.ListGroupItem("Item 3"),
+                dbc.ListGroupItem(
+                    [
+                        dbc.ListGroupItemHeading("Item 4 heading"),
+                        dbc.ListGroupItemText("Item 4 text"),
+                    ]
+                ),
+            ]
+        ),
+    ],
+    className="mb-4",
+)
