+++ docs/demos/theme_explorer/table.py
@@ -1,0 +1,52 @@
+import dash_bootstrap_components as dbc
+import dash_html_components as html
+
+from .util import make_subheading
+
+table = html.Div(
+    [
+        make_subheading("Table", "table"),
+        dbc.Table(
+            [
+                html.Thead(
+                    html.Tr(
+                        [
+                            html.Th("#"),
+                            html.Th("First name"),
+                            html.Th("Last name"),
+                        ]
+                    )
+                ),
+                html.Tbody(
+                    [
+                        html.Tr(
+                            [
+                                html.Th("1", scope="row"),
+                                html.Td("Tom"),
+                                html.Td("Cruise"),
+                            ]
+                        ),
+                        html.Tr(
+                            [
+                                html.Th("2", scope="row"),
+                                html.Td("Jodie"),
+                                html.Td("Foster"),
+                            ]
+                        ),
+                        html.Tr(
+                            [
+                                html.Th("3", scope="row"),
+                                html.Td("Chadwick"),
+                                html.Td("Boseman"),
+                            ]
+                        ),
+                    ]
+                ),
+            ],
+            responsive=True,
+            striped=True,
+            hover=True,
+        ),
+    ],
+    className="mb-4",
+)
