+++ docs/demos/theme_explorer/toast.py
@@ -1,0 +1,24 @@
+import dash_bootstrap_components as dbc
+import dash_html_components as html
+
+from .util import make_subheading
+
+toast = html.Div(
+    [
+        make_subheading("Toast", "toast"),
+        dbc.Button(
+            "Open toast",
+            id="auto-toast-toggle",
+            color="primary",
+            className="mb-3",
+        ),
+        dbc.Toast(
+            html.P("This is the content of the toast", className="mb-0"),
+            id="auto-toast",
+            header="This is the header",
+            icon="primary",
+            duration=4000,
+        ),
+    ],
+    className="mb-2",
+)
